refactor: Rank enum 클래스 패키지 변경
* 성격이 도메인과 관련있으므로 models로 변경
