+++ src/main/java/study/step4/models/Rank.java
@@ -1,4 +1,4 @@
-package study.step4;
+package study.step4.models;
 
 import java.util.Arrays;
