+++ src/test/java/study/step4/models/WinnersTest.java
@@ -1,7 +1,6 @@
 package study.step4.models;
 
 import org.junit.jupiter.api.Test;
-import study.step4.Rank;
 
 import java.util.Arrays;
